Implement saving of trajectories as json
Additionally, move the buddhabrot rendering portion and the PPM
interaction into seperate internal modules.
